+++ esp-wifi/src/timer/timer_esp32c2.rs
@@ -20,6 +20,14 @@ pub fn setup_radio_isr() {
     }
 }
 
+pub fn shutdown_radio_isr() {
+    #[cfg(feature = "ble")]
+    {
+        interrupt::disable(crate::hal::Cpu::ProCpu, Interrupt::LP_TIMER);
+        interrupt::disable(crate::hal::Cpu::ProCpu, Interrupt::BT_MAC);
+    }
+}
+
 #[cfg(feature = "wifi")]
 #[no_mangle]
 extern "C" fn WIFI_MAC() {