+++ esp-wifi/src/timer/xtensa.rs
@@ -1,6 +1,3 @@
-use core::cell::RefCell;
-
-use critical_section::Mutex;
 use esp_hal::interrupt::InterruptHandler;
 
 use crate::{
@@ -10,9 +7,10 @@ use crate::{
 };
 
 /// The timer responsible for time slicing.
-static TIMER1: Mutex<RefCell<Option<TimeBase>>> = Mutex::new(RefCell::new(None));
 const TIMESLICE_FREQUENCY: fugit::HertzU64 =
     fugit::HertzU64::from_raw(crate::CONFIG.tick_rate_hz as u64);
+
+use super::TIMER;
 
 // Time keeping
 pub const TICKS_PER_SECOND: u64 = 1_000_000;
@@ -31,8 +29,17 @@ pub fn setup_timer(mut timer1: TimeBase) -> Result<(), esp_hal::timer::Error> {
     timer1.start(TIMESLICE_FREQUENCY.into_duration())?;
     critical_section::with(|cs| {
         timer1.enable_interrupt(true);
-        TIMER1.borrow_ref_mut(cs).replace(timer1);
+        TIMER.borrow_ref_mut(cs).replace(timer1);
     });
+    Ok(())
+}
+
+pub fn disable_timer() -> Result<(), esp_hal::timer::Error> {
+    critical_section::with(|cs| {
+        unwrap!(TIMER.borrow_ref_mut(cs).as_mut()).enable_interrupt(false);
+        unwrap!(TIMER.borrow_ref_mut(cs).as_mut()).cancel().unwrap();
+    });
+
     Ok(())
 }
 
@@ -47,9 +54,15 @@ pub fn setup_multitasking() {
     }
 }
 
+pub fn disable_multitasking() {
+    xtensa_lx::interrupt::disable_mask(
+        1 << 29, // Disable Software1
+    );
+}
+
 fn do_task_switch(context: &mut TrapFrame) {
     critical_section::with(|cs| {
-        let mut timer = TIMER1.borrow_ref_mut(cs);
+        let mut timer = TIMER.borrow_ref_mut(cs);
         let timer = unwrap!(timer.as_mut());
         timer.clear_interrupt();
     });