+++ examples/src/bin/wifi_access_point.rs
@@ -53,10 +53,10 @@ fn main() -> ! {
     )
     .unwrap();
 
-    let wifi = peripherals.WIFI;
+    let mut wifi = peripherals.WIFI;
     let mut socket_set_entries: [SocketStorage; 3] = Default::default();
     let (iface, device, mut controller, sockets) =
-        create_network_interface(&init, wifi, WifiApDevice, &mut socket_set_entries).unwrap();
+        create_network_interface(&init, &mut wifi, WifiApDevice, &mut socket_set_entries).unwrap();
     let mut wifi_stack = WifiStack::new(iface, device, sockets, current_millis);
 
     let client_config = Configuration::AccessPoint(AccessPointConfiguration {